+++ src/headers.h
@@ -927,6 +927,7 @@ static const int OP_OFUN_OFFSETS[4] = { 0, 4, 8, 12 };
 #define KWS_CKORDERI 23
 #define KWS_BDIVWI 24
 #define KWS_CCOMPRESSI 25
+#define KWS_TREEBLASI 26
 
 #define CQ_TNOBSI 0
 #define CQ_ENOBSI   1